feat: moving the test suite over to using mocha+chai
