--- .alexignore
@@ -1,1 +1,0 @@
-CHANGELOG.md
