--- .alexrc
@@ -1,3 +1,0 @@
-{
-  "profanitySureness": 1
-}
