+++ .eslintignore
@@ -1,3 +1,3 @@
-node_modules/
+.nyc_output/
 coverage/
-!.alexrc.js
+node_modules/
