+++ .gitignore
@@ -1,2 +1,3 @@
-node_modules/
+.nyc_output/
 coverage/
+node_modules/
