+++ .mocharc.json
@@ -1,0 +1,5 @@
+{
+  "watch-files": ["src/*.js", "test/**/*.js"],
+  "recursive": true,
+  "reporter": "spec"
+}
