+++ package.json
@@ -10,16 +10,15 @@
     "url": "git://github.com/readmeio/oas-to-har.git"
   },
   "engines": {
-    "node": "^12 || ^14 || ^16"
+    "node": ">=14"
   },
   "scripts": {
     "lint": "eslint .",
-    "lint:docs": "alex . .github/",
     "prepare": "husky install",
-    "pretest": "npm run lint && npm run lint:docs",
+    "pretest": "npm run lint",
     "prettier": "prettier --list-different --write \"./**/**.{js}\"",
     "release": "npx conventional-changelog-cli -i CHANGELOG.md -s && git add CHANGELOG.md",
-    "test": "jest --coverage"
+    "test": "nyc mocha"
   },
   "dependencies": {
     "@readme/oas-extensions": "^14.2.0",
@@ -33,21 +32,16 @@
     "@commitlint/config-conventional": "^16.0.0",
     "@readme/eslint-config": "^8.5.1",
     "@readme/oas-examples": "^5.0.0",
-    "alex": "^10.0.0",
+    "chai": "^4.3.6",
     "datauri": "^4.1.0",
     "eslint": "^8.11.0",
+    "har-validator": "^5.1.5",
     "husky": "^7.0.4",
-    "jest": "^27.4.7",
-    "jest-expect-har": "^3.0.1",
+    "mocha": "^9.2.2",
+    "nyc": "^15.1.0",
     "prettier": "^2.6.0"
   },
   "prettier": "@readme/eslint-config/prettier",
-  "jest": {
-    "testPathIgnorePatterns": [
-      "/__tests__/__fixtures__/",
-      "/__tests__/(.*)/__fixtures__/"
-    ]
-  },
   "commitlint": {
     "extends": [
       "@commitlint/config-conventional"
