+++ test/.eslintrc
@@ -1,0 +1,3 @@
+{
+  "extends": "@readme/eslint-config/testing-mocha"
+}
